fastapi_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
